# InitializeBuffer: use maximum 2 h2d copies