This provides a significant speedup when autotuning, where buffers are repeatedly initialized.
Instead of repeatedly copying the same data to different locations on the device, it is copied once (using up to two host to device copies) and then replicated on the device using a custom kernel.

Using the paxml container from JAX-Toolbox and the 5B configuration on H100, the total runtime of the `gemm-algorithm-picker` pass for the main JITed function decreased from around 4.8s to 0.16s. With the same model and `--xla_gpu_triton_gemm_any=true` to make sure the Triton autotuner does a lot of work, that pass speeds up from 17.2s to 14.2s with this change.